+++ packages/components/button-group/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.8](https://github.com/Synerise/synerise-design/compare/@synerise/ds-button-group@0.1.7...@synerise/ds-button-group@0.1.8) (2019-11-09)
+
+**Note:** Version bump only for package @synerise/ds-button-group
+
 ## [0.1.7](https://github.com/Synerise/synerise-design/compare/@synerise/ds-button-group@0.1.6...@synerise/ds-button-group@0.1.7) (2019-11-06)
 
 **Note:** Version bump only for package @synerise/ds-button-group