+++ packages/components/core/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.7](https://github.com/synerise/synerise-design/compare/@synerise/ds-core@0.6.6...@synerise/ds-core@0.6.7) (2019-11-09)
+
+**Note:** Version bump only for package @synerise/ds-core
+
 ## [0.6.6](https://github.com/synerise/synerise-design/compare/@synerise/ds-core@0.6.5...@synerise/ds-core@0.6.6) (2019-11-06)
 
 **Note:** Version bump only for package @synerise/ds-core