+++ packages/components/page-header/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.1](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.5.0...@synerise/ds-page-header@0.5.1) (2019-11-09)
+
+**Note:** Version bump only for package @synerise/ds-page-header
+
 # [0.5.0](https://github.com/synerise/synerise-design/compare/@synerise/ds-page-header@0.4.6...@synerise/ds-page-header@0.5.0) (2019-11-06)
 
 ### Features