+++ packages/components/progress-bar/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.7](https://github.com/Synerise/synerise-design/compare/@synerise/ds-progress-bar@0.1.6...@synerise/ds-progress-bar@0.1.7) (2019-11-09)
+
+**Note:** Version bump only for package @synerise/ds-progress-bar
+
 ## [0.1.6](https://github.com/Synerise/synerise-design/compare/@synerise/ds-progress-bar@0.1.5...@synerise/ds-progress-bar@0.1.6) (2019-11-06)
 
 **Note:** Version bump only for package @synerise/ds-progress-bar